SC-7998 move version route to be a feathers service
