+++ CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 Allowed Types of change: `Added`, `Changed`, `Deprecated`, `Removed`, `Fixed`, `Security`
 
+## [25.3.4]
+
+### Changed
+
+- SC-7998 - use default service setup for /version
+
 ## [25.3.3] (pick from 25.2)
 
 ### Removed
