+++ src/services/docs/openapi.yaml
@@ -1,8 +1,7 @@
 info:
   title: Schul-Cloud Legacy API
-  description:
-    This is the API specification for the Schul-Cloud legacy API.
-    
+  description: This is the API specification for the Schul-Cloud legacy API.
+
   contact:
     name: support
     email: info@hpi-schul-cloud.de
@@ -17,6 +16,6 @@ paths:
   /me:
     $ref: '../me/docs/openapi.yaml#/paths/~1me'
   /version:
-    $ref: '../versionService/openapi.yaml#/paths/~1version'
+    $ref: '../versionService/docs/openapi.yaml#/paths/~1version'
 
 openapi: 3.0.2
