+++ src/services/index.js
@@ -50,6 +50,7 @@ const etherpad = require('./etherpad');
 const storageProvider = require('./storageProvider');
 const activation = require('./activation');
 const config = require('./config');
+const version = require('./versionService');
 const docs = require('./docs');
 
 module.exports = function initializeServices() {
@@ -108,6 +109,7 @@ module.exports = function initializeServices() {
 	app.configure(storageProvider);
 	app.configure(activation);
 	app.configure(config);
+	app.configure(version);
 
 	// initialize events
 	newsEvents.configure(app);
