+++ src/services/versionService/hooks/index.js
@@ -1,0 +1,24 @@
+const before = {
+	all: [],
+	find: [],
+	get: [],
+	create: [],
+	update: [],
+	patch: [],
+	remove: [],
+};
+
+const after = {
+	all: [],
+	find: [],
+	get: [],
+	create: [],
+	update: [],
+	patch: [],
+	remove: [],
+};
+
+module.exports = {
+	before,
+	after,
+};
